refactor: resetup project structure

+++ src/database/UserData.cpp
@@ -1,11 +1,11 @@
 #include <UserData.hpp>
 
-const char* UserData::gameModeNames[GameMode::GameModeCount] = {"Easy", "Medium", "Hard", "Insane", "Rain Day", "Endless"};
+const char* UserData::gameModeNames[(int)GameMode::GameModeCount] = {"Easy", "Medium", "Hard", "Insane", "Rain Day", "Endless"};
 
 UserData::UserData()
     : username(""), password("")
 {
-    for (int i = 0; i < GameMode::GameModeCount; ++i) {
+    for (int i = 0; i < (int)GameMode::GameModeCount; ++i) {
         highscore.at(i) = 0;
     }
 }
@@ -13,18 +13,18 @@ UserData::UserData()
 UserData::UserData(std::string username, std::string password, int easy, int medium, int hard, int insane, int rainday, int endless)
     : username(username), password(password)
 {
-    highscore.at(GameMode::easy) = easy;
-    highscore.at(GameMode::medium) = medium;
-    highscore.at(GameMode::hard) = hard;
-    highscore.at(GameMode::insane) = insane;
-    highscore.at(GameMode::rainday) = rainday;
-    highscore.at(GameMode::endless) = endless;
+    highscore.at((int)GameMode::easy) = easy;
+    highscore.at((int)GameMode::medium) = medium;
+    highscore.at((int)GameMode::hard) = hard;
+    highscore.at((int)GameMode::insane) = insane;
+    highscore.at((int)GameMode::rainday) = rainday;
+    highscore.at((int)GameMode::endless) = endless;
 }
 
 UserData::UserData(std::string username, std::string password)
     : username(username), password(password)
 {
-    for (int i = 0; i < GameMode::GameModeCount; ++i) {
+    for (int i = 0; i < (int)GameMode::GameModeCount; ++i) {
         highscore.at(i) = 0;
     }
 }
@@ -44,14 +44,14 @@ std::string UserData::getPassword() const
     return password;
 }
 
-std::array<int, UserData::GameMode::GameModeCount> UserData::getHighscore() const
+std::array<int, (int)UserData::GameMode::GameModeCount> UserData::getHighscore() const
 {
     return highscore;
 }
 
 int UserData::getHighscore(GameMode mode) const
 {
-    return highscore.at(mode);
+    return highscore.at((int)mode);
 }
 
 void UserData::setPassword(const std::string &newPassword)
@@ -61,13 +61,13 @@ void UserData::setPassword(const std::string &newPassword)
 
 void UserData::setHighscore(int newHighscore, GameMode gameMode)
 {
-    highscore.at(gameMode) = newHighscore;
+    highscore.at((int)gameMode) = newHighscore;
 }
 
 void UserData::updateHighscore(int newHighscore, GameMode gameMode)
 {
-    if (newHighscore > highscore.at(gameMode)) {
-        highscore.at(gameMode) = newHighscore;
+    if (newHighscore > highscore.at((int)gameMode)) {
+        highscore.at((int)gameMode) = newHighscore;
     }
 }
